+++ app-demo/src/main/java/mobi/lab/components/compose/demo/NavDestination.kt
@@ -20,4 +20,7 @@ object NavDestination {
 
     @Serializable
     object TextField
+
+    @Serializable
+    object TopAppBar
 }