+++ app-demo/src/main/java/mobi/lab/components/compose/demo/textfield/TextFieldDestination.kt
@@ -16,6 +16,7 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.filled.Info
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Text
@@ -57,7 +58,11 @@ fun TextFieldDestination(onNavigateUp: () -> Unit, onToggleLightDarkModeClicked:
             topBar = {
                 LabTopAppBar(
                     stringResource(R.string.title_text_fields),
-                    navConfig = upNavConfig(onNavigateUp),
+                    navConfig = upNavConfig(
+                        icon = ImageSource.vector(Icons.AutoMirrored.Filled.ArrowBack),
+                        contentDescription = stringResource(R.string.btn_back),
+                        onClick = onNavigateUp
+                    ),
                     actions = { LightDarkModeMenu(onToggleLightDarkModeClicked) }
                 )
             }