+++ lib/src/main/java/mobi/lab/components/compose/widget/progress/LabIndeterminateCircularIndicatorDefaults.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
 import mobi.lab.components.compose.theme.LabTheme
 
-public object LabIndeterminateProgressDefaults {
+public object LabIndeterminateCircularIndicatorDefaults {
 
     @Composable
     public fun defaultIndeterminateProgressColor(): Color {