+++ lib/src/main/java/mobi/lab/components/compose/widget/topappbar/LabTopAppBar.kt
@@ -2,32 +2,27 @@
 
 package mobi.lab.components.compose.widget.topappbar
 
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.RowScope
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.FavoriteBorder
+import androidx.compose.material.icons.filled.Call
+import androidx.compose.material.icons.filled.Edit
+import androidx.compose.material.icons.filled.Email
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.IconButtonDefaults
+import androidx.compose.material3.LargeTopAppBar
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarColors
+import androidx.compose.material3.TopAppBarScrollBehavior
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import mobi.lab.components.compose.R
-import mobi.lab.components.compose.theme.LabTheme
+import androidx.compose.ui.unit.Dp
 import mobi.lab.components.compose.util.PreviewContainer
 import mobi.lab.components.compose.widget.button.LabButtonDefaults
 import mobi.lab.components.compose.widget.button.LabIconButton
-import mobi.lab.components.compose.widget.image.IconFromSource
 import mobi.lab.components.compose.widget.image.ImageSource
 
 @Composable
@@ -35,35 +30,14 @@ public fun LabTopAppBar(
     title: String?,
     modifier: Modifier = Modifier,
     navConfig: NavConfig? = null,
-    titleStyle: TextStyle = LabTheme.typography.titleLarge,
-    colors: TopAppBarColors = LabTopAppBarDefaults.colors(),
+    titleStyle: TextStyle = LabTopAppBarDefaults.topAppBarTitleStyle,
+    colors: TopAppBarColors = LabTopAppBarDefaults.topAppBarColors(),
     actions: @Composable RowScope.() -> Unit = {},
+    expandedHeight: Dp = LabTopAppBarDefaults.topAppBarExpandedHeight,
+    windowInsets: WindowInsets = LabTopAppBarDefaults.topAppBarWindowInsets,
+    scrollBehavior: TopAppBarScrollBehavior? = null,
 ) {
-    // We will not use the MyToolbar as the left text button is not any official spec
-    TopAppBar(
-        navigationIcon = {
-            if (navConfig != null) {
-                // Up navigation and left buttons can't be used at the same time.
-                Row {
-                    IconButton(
-                        onClick = navConfig.onClick,
-                        colors = IconButtonDefaults.iconButtonColors().copy(
-                            contentColor = colors.navigationIconContentColor,
-                            containerColor = colors.containerColor,
-                        )
-                    ) {
-                        IconFromSource(
-                            source = navConfig.icon,
-                            contentDescription = stringResource(R.string.lab_back),
-                        )
-                    }
-                    Spacer(Modifier.size(16.dp))
-                }
-            }
-        },
-        actions = actions,
-        modifier = modifier,
-        colors = colors,
+    LabTopAppBar(
         title = {
             if (title != null) {
                 Text(
@@ -74,41 +48,145 @@ public fun LabTopAppBar(
                     text = title,
                 )
             }
-        }
+        },
+        modifier = modifier,
+        navigationIcon = {
+            if (navConfig != null) {
+                // In the design system template, this here is an instance of the icon button
+                // as per design system owner's.
+                LabIconButton(
+                    icon = navConfig.icon,
+                    contentDescription = navConfig.contentDescription,
+                    onClick = navConfig.onClick,
+                    colors = LabButtonDefaults.iconButtonColors().copy(
+                        containerColor = colors.containerColor,
+                        contentColor = colors.navigationIconContentColor
+                    )
+                )
+            }
+        },
+        actions = actions,
+        expandedHeight = expandedHeight,
+        windowInsets = windowInsets,
+        colors = colors,
+        scrollBehavior = scrollBehavior,
     )
 }
 
-public object LabTopAppBarDefaults {
-    @Composable
-    public fun colors(): TopAppBarColors {
-        return TopAppBarColors(
-            // We only need this one atm:
-            containerColor = LabTheme.colors.background,
-            scrolledContainerColor = LabTheme.colors.background,
-            navigationIconContentColor = LabTheme.colors.primary,
-            titleContentColor = LabTheme.colors.onBackground,
-            actionIconContentColor = LabTheme.colors.primary,
-        )
-    }
+@Composable
+public fun LabTopAppBar(
+    title: @Composable () -> Unit,
+    modifier: Modifier = Modifier,
+    navigationIcon: @Composable () -> Unit = {},
+    actions: @Composable RowScope.() -> Unit = {},
+    expandedHeight: Dp = LabTopAppBarDefaults.topAppBarExpandedHeight,
+    windowInsets: WindowInsets = LabTopAppBarDefaults.topAppBarWindowInsets,
+    colors: TopAppBarColors = LabTopAppBarDefaults.topAppBarColors(),
+    scrollBehavior: TopAppBarScrollBehavior? = null
+) {
+    TopAppBar(
+        title = title,
+        modifier = modifier,
+        navigationIcon = navigationIcon,
+        actions = actions,
+        expandedHeight = expandedHeight,
+        windowInsets = windowInsets,
+        colors = colors,
+        scrollBehavior = scrollBehavior,
+    )
 }
 
-@Preview(showBackground = true, showSystemUi = true)
 @Composable
-private fun LabTopAppBarWithButtonComposableWithUpNavPreview() {
-    PreviewContainer(Modifier.height(200.dp)) {
+public fun LabLargeTopAppBar(
+    title: String?,
+    modifier: Modifier = Modifier,
+    navConfig: NavConfig? = null,
+    titleStyle: TextStyle = LabTopAppBarDefaults.largeTopAppBarTitleStyle,
+    colors: TopAppBarColors = LabTopAppBarDefaults.largeTopAppBarColors(),
+    actions: @Composable RowScope.() -> Unit = {},
+    collapsedHeight: Dp = LabTopAppBarDefaults.largeAppBarCollapsedHeight,
+    expandedHeight: Dp = LabTopAppBarDefaults.largeAppBarExpandedHeight,
+    windowInsets: WindowInsets = LabTopAppBarDefaults.largeTopAppBarWindowInsets,
+    scrollBehavior: TopAppBarScrollBehavior? = null,
+) {
+    LabLargeTopAppBar(
+        title = {
+            if (title != null) {
+                Text(
+                    color = colors.titleContentColor,
+                    style = titleStyle,
+                    maxLines = 1,
+                    overflow = TextOverflow.Ellipsis,
+                    text = title,
+                )
+            }
+        },
+        modifier = modifier,
+        navigationIcon = {
+            if (navConfig != null) {
+                // In the design system template, this here is an instance of the icon button
+                // as per design system owner's.
+                LabIconButton(
+                    icon = navConfig.icon,
+                    contentDescription = navConfig.contentDescription,
+                    onClick = navConfig.onClick,
+                    colors = LabButtonDefaults.iconButtonColors().copy(
+                        containerColor = colors.containerColor,
+                        contentColor = colors.navigationIconContentColor
+                    )
+                )
+            }
+        },
+        actions = actions,
+        collapsedHeight = collapsedHeight,
+        expandedHeight = expandedHeight,
+        windowInsets = windowInsets,
+        colors = colors,
+        scrollBehavior = scrollBehavior,
+    )
+}
+
+@Composable
+public fun LabLargeTopAppBar(
+    title: @Composable () -> Unit,
+    modifier: Modifier = Modifier,
+    navigationIcon: @Composable () -> Unit = {},
+    actions: @Composable RowScope.() -> Unit = {},
+    collapsedHeight: Dp = LabTopAppBarDefaults.largeAppBarCollapsedHeight,
+    expandedHeight: Dp = LabTopAppBarDefaults.largeAppBarExpandedHeight,
+    windowInsets: WindowInsets = LabTopAppBarDefaults.largeTopAppBarWindowInsets,
+    colors: TopAppBarColors = LabTopAppBarDefaults.largeTopAppBarColors(),
+    scrollBehavior: TopAppBarScrollBehavior? = null
+) {
+    LargeTopAppBar(
+        title = title,
+        modifier = modifier,
+        navigationIcon = navigationIcon,
+        actions = actions,
+        collapsedHeight = collapsedHeight,
+        expandedHeight = expandedHeight,
+        windowInsets = windowInsets,
+        colors = colors,
+        scrollBehavior = scrollBehavior,
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun LabTopAppBarLightWithTitle() {
+    PreviewContainer {
         LabTopAppBar(
-            navConfig = upNavConfig(),
             title = "Title"
         )
     }
 }
 
-@Preview(showBackground = true, showSystemUi = true)
+@Preview(showBackground = true)
 @Composable
-private fun LabTopAppBarWithButtonComposableWithNoNavConfigPreview() {
-    PreviewContainer(Modifier.height(200.dp)) {
+private fun LabTopAppBarLightWithBackAndTitle() {
+    PreviewContainer {
         LabTopAppBar(
-            navConfig = null,
+            navConfig = upNavConfig(),
             title = "Title"
         )
     }
@@ -117,20 +195,186 @@ private fun LabTopAppBarWithButtonComposableWithNoNavConfigPreview() {
 /**
  * For now, actions need to be specified via a Composable.
  */
-@Preview(showBackground = true, showSystemUi = true)
+@Preview(showBackground = true)
 @Composable
-private fun LabTopAppBarWithButtonComposableWithRightButtonPreview() {
-    PreviewContainer(Modifier.height(200.dp)) {
+private fun LabTopAppBarLightWithBackAndTitleAndAction() {
+    PreviewContainer {
         LabTopAppBar(
-            navConfig = null,
+            navConfig = upNavConfig(),
             title = "Title",
             actions = {
                 LabIconButton(
-                    icon = ImageSource.vector(Icons.Filled.FavoriteBorder),
-                    contentDescription = "Like",
+                    icon = ImageSource.vector(Icons.Filled.Edit),
+                    contentDescription = "Edit",
                     onClick = {},
-                    enabled = true,
-                    colors = LabButtonDefaults.iconButtonColors().copy(containerColor = LabTheme.colors.background)
+                )
+                LabIconButton(
+                    icon = ImageSource.vector(Icons.Filled.Email),
+                    contentDescription = "Email",
+                    onClick = {},
+                )
+                LabIconButton(
+                    icon = ImageSource.vector(Icons.Filled.Call),
+                    contentDescription = "Call",
+                    onClick = {},
+                )
+            },
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun LabTopAppBarDarkWithTitle() {
+    PreviewContainer(isDark = true) {
+        LabTopAppBar(
+            navConfig = null,
+            title = "Title"
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun LabTopAppBarDarkWithBackAndTitle() {
+    PreviewContainer(isDark = true) {
+        LabTopAppBar(
+            navConfig = upNavConfig(),
+            title = "Title"
+        )
+    }
+}
+
+/**
+ * For now, actions need to be specified via a Composable.
+ */
+@Preview(showBackground = true)
+@Composable
+private fun LabTopAppBarDarkWithBackAndTitleAndAction() {
+    PreviewContainer(isDark = true) {
+        LabTopAppBar(
+            navConfig = upNavConfig(),
+            title = "Title",
+            actions = {
+                LabIconButton(
+                    icon = ImageSource.vector(Icons.Filled.Edit),
+                    contentDescription = "Edit",
+                    onClick = {},
+                )
+                LabIconButton(
+                    icon = ImageSource.vector(Icons.Filled.Email),
+                    contentDescription = "Email",
+                    onClick = {},
+                )
+                LabIconButton(
+                    icon = ImageSource.vector(Icons.Filled.Call),
+                    contentDescription = "Call",
+                    onClick = {},
+                )
+            },
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun LabLargeTopAppBarLightWithTitle() {
+    PreviewContainer {
+        LabLargeTopAppBar(
+            title = "Title"
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun LabLargeTopAppBarLightWithBackAndTitle() {
+    PreviewContainer {
+        LabLargeTopAppBar(
+            navConfig = upNavConfig(),
+            title = "Title"
+        )
+    }
+}
+
+/**
+ * For now, actions need to be specified via a Composable.
+ */
+@Preview(showBackground = true)
+@Composable
+private fun LabLargeTopAppBarLightWithBackAndTitleAndAction() {
+    PreviewContainer {
+        LabLargeTopAppBar(
+            navConfig = upNavConfig(),
+            title = "Title",
+            actions = {
+                LabIconButton(
+                    icon = ImageSource.vector(Icons.Filled.Edit),
+                    contentDescription = "Edit",
+                    onClick = {},
+                )
+                LabIconButton(
+                    icon = ImageSource.vector(Icons.Filled.Email),
+                    contentDescription = "Email",
+                    onClick = {},
+                )
+                LabIconButton(
+                    icon = ImageSource.vector(Icons.Filled.Call),
+                    contentDescription = "Call",
+                    onClick = {},
+                )
+            },
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun LabLargeTopAppBarDarkWithTitle() {
+    PreviewContainer(isDark = true) {
+        LabLargeTopAppBar(
+            navConfig = null,
+            title = "Title"
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun LabLargeTopAppBarDarkWithBackAndTitle() {
+    PreviewContainer(isDark = true) {
+        LabLargeTopAppBar(
+            navConfig = upNavConfig(),
+            title = "Title"
+        )
+    }
+}
+
+/**
+ * For now, actions need to be specified via a Composable.
+ */
+@Preview(showBackground = true)
+@Composable
+private fun LabLargeTopAppBarDarkWithBackAndTitleAndAction() {
+    PreviewContainer(isDark = true) {
+        LabLargeTopAppBar(
+            navConfig = upNavConfig(),
+            title = "Title",
+            actions = {
+                LabIconButton(
+                    icon = ImageSource.vector(Icons.Filled.Edit),
+                    contentDescription = "Edit",
+                    onClick = {},
+                )
+                LabIconButton(
+                    icon = ImageSource.vector(Icons.Filled.Email),
+                    contentDescription = "Email",
+                    onClick = {},
+                )
+                LabIconButton(
+                    icon = ImageSource.vector(Icons.Filled.Call),
+                    contentDescription = "Call",
+                    onClick = {},
                 )
             },
         )